Fix most code review issues
