+++ packages/backend/src/api/controllers/SearchController.ts
@@ -13,7 +13,7 @@ export class SearchController {
   ) {}
 
   async getSearchRedirect(query: string): Promise<ControllerResult> {
-    const parsed = parseSearchQuery(query)
+    const parsed = parseSearchQuery(query.trim())
 
     let response: ControllerResult | undefined
 
@@ -21,16 +21,24 @@ export class SearchController {
       response = await this.searchForEthereumAddress(parsed.ethereumAddress)
     }
 
-    if (!response && parsed.stateTreeHash) {
-      response = await this.searchForRootHash(parsed.stateTreeHash)
-    }
-
     if (!response && parsed.starkKey) {
       response = await this.searchForStarkKey(parsed.starkKey)
     }
 
+    if (!response && parsed.positionOrVaultId) {
+      response = await this.searchForPositionOrVaultId(parsed.positionOrVaultId)
+    }
+
+    if (!response && parsed.stateUpdateId) {
+      response = await this.searchForStateUpdateId(parsed.stateUpdateId)
+    }
+
+    if (!response && parsed.stateUpdateRootHash) {
+      response = await this.searchForRootHash(parsed.stateUpdateRootHash)
+    }
+
     if (!response) {
-      const content = "Search query couldn't be found"
+      const content = 'No results found for your query'
       return { type: 'not found', content }
     }
 
@@ -48,14 +56,9 @@ export class SearchController {
       return
     }
 
-    const positionId = await this.positionRepository.findIdByStarkKey(
-      userRegistrationEvent.starkKey
-    )
-    if (positionId === undefined) {
-      return
-    }
+    const starkKey = userRegistrationEvent.starkKey.toString()
 
-    return { type: 'redirect', url: `/positions/${positionId}` }
+    return { type: 'redirect', url: `/users/${starkKey}` }
   }
 
   private async searchForRootHash(
@@ -80,29 +83,53 @@ export class SearchController {
       return
     }
 
-    const positionId = await this.positionRepository.findIdByStarkKey(
-      userRegistrationEvent.starkKey
-    )
-    if (positionId === undefined) {
+    return { type: 'redirect', url: `/users/${starkKey.toString()}` }
+  }
+
+  private async searchForPositionOrVaultId(
+    positionId: bigint
+  ): Promise<ControllerResult | undefined> {
+    const position = await this.positionRepository.findById(positionId)
+    if (position === undefined) {
       return
     }
 
-    return { type: 'redirect', url: `/positions/${positionId}` }
+    const starkKey = position.starkKey.toString()
+
+    return { type: 'redirect', url: `/user/${starkKey}` }
+  }
+
+  private async searchForStateUpdateId(
+    stateUpdateId: number
+  ): Promise<ControllerResult | undefined> {
+    const stateUpdate = await this.stateUpdateRepository.findById(stateUpdateId)
+    if (stateUpdate === undefined) {
+      return
+    }
+
+    return { type: 'redirect', url: `/state-updates/${stateUpdateId}` }
   }
 }
 
 interface ParsedQuery {
   ethereumAddress?: EthereumAddress
-  stateTreeHash?: PedersenHash
   starkKey?: StarkKey
+  positionOrVaultId?: bigint
+  stateUpdateId?: number
+  stateUpdateRootHash?: PedersenHash
 }
 
 export function parseSearchQuery(query: string): ParsedQuery {
   const parsed: ParsedQuery = {}
-  parsed.stateTreeHash = tryOrUndefined(() => PedersenHash(query))
   parsed.starkKey = tryOrUndefined(() => StarkKey(query))
   parsed.ethereumAddress = tryOrUndefined(() => EthereumAddress(query))
-
+  parsed.positionOrVaultId = query.startsWith('#')
+    ? BigInt(query.slice(1))
+    : undefined
+  parsed.stateUpdateId = query.startsWith('@')
+    ? parseInt(query.slice(1))
+    : undefined
+  parsed.stateUpdateRootHash = tryOrUndefined(() => PedersenHash(query))
   return parsed
 }
