+++ packages/backend/src/api/routers/OldFrontendRouter.ts
@@ -12,9 +12,9 @@ import { z } from 'zod'
 import { ForcedTradeOfferController } from '../controllers/ForcedTradeOfferController'
 import { ForcedTransactionController } from '../controllers/ForcedTransactionController'
 import { OldHomeController } from '../controllers/OldHomeController'
+import { OldSearchController } from '../controllers/OldSearchController'
 import { OldStateUpdateController } from '../controllers/OldStateUpdateController'
 import { PositionController } from '../controllers/PositionController'
-import { SearchController } from '../controllers/SearchController'
 import { withTypedContext } from './types'
 import { applyControllerResult } from './utils'
 
@@ -24,7 +24,7 @@ export function createOldFrontendRouter(
   forcedTradeOfferController: ForcedTradeOfferController,
   forcedTransactionController: ForcedTransactionController,
   stateUpdateController: OldStateUpdateController,
-  searchController: SearchController
+  searchController: OldSearchController
 ) {
   const router = new Router()
